features/workflow: fix Windows test failures with polling helper (Issue #309)
Tests used fixed time.Sleep() delays to wait for async workflow completion.
On slower systems (Windows CI), workflows took longer than sleep duration,
causing tests to check status before workflows finished, leading to test
failures with 'pending' status.

Root cause analysis:
- Workflow engine spawns async goroutines: go e.executeWorkflowAsync()
- Tests used time.Sleep(100-200ms) to wait for completion
- Windows CI slower execution meant workflows still running after sleep
- Tests checked status before terminal state reached

Solution implemented:
Created shared waitForWorkflowCompletion() helper in test_helpers.go that:
- Polls execution.GetStatus() every 10ms instead of fixed wait
- Returns immediately when workflow reaches terminal state
- Times out with detailed error after specified duration
- Uses GetStatus() method for proper locking

Changes:
- features/workflow/test_helpers.go (NEW): Created shared polling helper
- 12 test files: Replaced 50+ time.Sleep() with helper or reverted incorrect replacements
- Pattern recognition for three delay types: workflow waits, polling delays, setup delays

Test results (measured):
- All workflow tests: PASS (4.850s runtime)
- Full test suite: 80 packages, 0 failures
- All validation gates: PASS

Fixes #309
